Factored CNNBLOCK from pix2pix discriminator
